\name{crypto_data_daily}
\alias{crypto_data_daily}
\docType{data}
\title{
Cryptocurrency Data (Daily)
}
\description{
Times-Series Data for 4 Cryptocurrencies. Bitcoin (BTC), Cardano (ADA), Ethereum (ETH) and Ripple(XRP) from 2018 until 2022.
}
\usage{data("crypto_data_daily")}
\format{
  A data frame with 5801 observations on the following 11 variables.
}

\examples{
data(crypto_data_daily)
## maybe str(crypto_data_daily) ; plot(crypto_data_daily) ...
}
\keyword{datasets}
