\name{crypto_data_hourly}
\alias{crypto_data_hourly}
\docType{data}
\title{
Cryptocurrency Data (hourly)
}
\description{
Times-Series Data for 4 Cryptocurrencies. Bitcoin (BTC), Cardano (ADA), Ethereum (ETH) and Ripple(XRP) from 2018 until 2022.
}
\usage{data("crypto_data_hourly")}
\format{
  A data frame with 19847 observations on the following 11 variables.
}

\examples{
data(crypto_data_hourly)
## maybe str(crypto_data_hourly) ; plot(crypto_data_hourly) ...
}
\keyword{datasets}
